+++ src/genesys/genesys-source-api.ts
@@ -2,6 +2,8 @@ import { validateNonNull } from '../utils/validate-non-null.js';
 import { GenesysApi } from './genesys-api.js';
 import { GenesysSourceConfig } from './model/genesys-source-config.js';
 import { removeTrailingSlash } from '../utils/remove-trailing-slash.js';
+import { fetchSourceResource, RequestInit } from '../utils/web-client.js';
+import { EntityType } from '../model';
 
 export class GenesysSourceApi extends GenesysApi {
   private config: GenesysSourceConfig = {};
@@ -55,7 +57,11 @@ export class GenesysSourceApi extends GenesysApi {
     return this.config.genesysSourceKnowledgeBaseId!;
   }
 
-  protected getUserAgent(): string | undefined {
-    return this.config.sourceUserAgent;
+  protected innerFetch<T>(
+    url: string,
+    init?: RequestInit,
+    entityName?: EntityType,
+  ): Promise<T> {
+    return fetchSourceResource(url, init, entityName);
   }
 }