+++ src/genesys/model/genesys-source-config.ts
@@ -1,7 +1,6 @@
 import { Config } from '../../config.js';
 
 export interface GenesysSourceConfig extends Config {
-  sourceUserAgent?: string;
   genesysSourceLoginUrl?: string;
   genesysSourceBaseUrl?: string;
   genesysSourceClientId?: string;