+++ src/salesforce/model/salesforce-config.ts
@@ -1,7 +1,6 @@
 import { Config } from '../../config.js';
 
 export interface SalesforceConfig extends Config {
-  sourceUserAgent?: string;
   salesforceLoginUrl?: string;
   salesforceBaseUrl?: string;
   salesforceApiVersion?: string;