+++ src/servicenow/servicenow-api.ts
@@ -1,6 +1,6 @@
 import { AuthenticationType, ServiceNowConfig } from './model/servicenow-config.js';
 import { ServiceNowArticle } from './model/servicenow-article.js';
-import { fetchResource } from '../utils/web-client.js';
+import { fetchSourceResource } from '../utils/web-client.js';
 import { ServiceNowArticleResponse } from './model/servicenow-article-response.js';
 import { ServiceNowArticleAttachment } from './model/servicenow-article-attachment.js';
 import { removeTrailingSlash } from '../utils/remove-trailing-slash.js';
@@ -119,12 +119,12 @@ export class ServiceNowApi {
   ): Promise<ServiceNowArticleAttachment> {
     const url = `${this.baseUrl}/api/now/attachment/${attachmentId}`;
     const requestInit = await this.buildRequestInit();
-    return fetchResource(url, requestInit, EntityType.DOCUMENT);
+    return fetchSourceResource(url, requestInit, EntityType.DOCUMENT);
   }
 
   public async downloadAttachment(url: string): Promise<Blob> {
     const requestInit = await this.buildRequestInit();
-    return fetchResource<Blob>(
+    return fetchSourceResource<Blob>(
       url,
       requestInit,
       EntityType.DOCUMENT,
@@ -139,7 +139,7 @@ export class ServiceNowApi {
   public async getArticle(id: string): Promise<ServiceNowSingleArticle | null> {
     const url = `${this.baseUrl}/api/sn_km_api/knowledge/articles/${id}?fields=kb_category,kb_knowledge_base,workflow_state,active,sys_updated_on,valid_to`;
     const requestInit = await this.buildRequestInit();
-    const json = await fetchResource<ServiceNowSingleArticleResponse>(
+    const json = await fetchSourceResource<ServiceNowSingleArticleResponse>(
       url,
       requestInit,
       EntityType.DOCUMENT,
@@ -188,7 +188,7 @@ export class ServiceNowApi {
 
     const requestInit = await this.buildRequestInit();
 
-    const json = await fetchResource<ServiceNowArticleResponse>(
+    const json = await fetchSourceResource<ServiceNowArticleResponse>(
       url,
       requestInit,
       EntityType.DOCUMENT,
@@ -218,7 +218,7 @@ export class ServiceNowApi {
 
     const requestInit = await this.buildRequestInit();
 
-    const json = await fetchResource<ServiceNowCategoryResponse>(
+    const json = await fetchSourceResource<ServiceNowCategoryResponse>(
       url,
       requestInit,
       EntityType.CATEGORY,
@@ -238,28 +238,30 @@ export class ServiceNowApi {
   }
 
   private async buildRequestInit(): Promise<RequestInit> {
-    const headers: Record<string, string> = {};
-
-    if (this.config.sourceUserAgent) {
-      headers['User-Agent'] = this.config.sourceUserAgent;
-    }
-
     if (this.isOAuth && this.oAuthToken.bearerToken) {
       if (this.isTokenExpired()) {
         await this.refreshAccessToken();
       }
 
-      headers['Authorization'] = `Bearer ${this.oAuthToken.bearerToken}`;
-    } else {
-      headers['Authorization'] =
-        'Basic ' +
-        Buffer.from(
-          this.config.servicenowUsername + ':' + this.config.servicenowPassword,
-          'utf-8',
-        ).toString('base64');
+      return {
+        headers: {
+          Authorization: `Bearer ${this.oAuthToken.bearerToken}`,
+        },
+      };
     }
 
-    return { headers };
+    return {
+      headers: {
+        Authorization:
+          'Basic ' +
+          Buffer.from(
+            this.config.servicenowUsername +
+              ':' +
+              this.config.servicenowPassword,
+            'utf-8',
+          ).toString('base64'),
+      },
+    };
   }
 
   private queryParams(): string {
@@ -378,23 +380,16 @@ export class ServiceNowApi {
 
   private async getAccessToken(bodyParams: URLSearchParams): Promise<void> {
     const url = `${this.baseUrl}/oauth_token.do`;
-
-    const headers: Record<string, string> = {
-      'Content-Type': 'application/x-www-form-urlencoded',
-    }
-
-    if (this.config.sourceUserAgent) {
-      headers['User-Agent'] = this.config.sourceUserAgent;
-    }
-
     const request = {
       method: 'POST',
       body: bodyParams,
-      headers,
+      headers: {
+        'Content-Type': 'application/x-www-form-urlencoded',
+      },
     };
 
     try {
-      const data = await fetchResource<ServiceNowAccessTokenResponse>(
+      const data = await fetchSourceResource<ServiceNowAccessTokenResponse>(
         url,
         request,
         undefined,
@@ -408,8 +403,8 @@ export class ServiceNowApi {
       this.oAuthToken = {
         bearerToken: data.access_token,
         refreshToken: data.refresh_token,
-        expiresAt: Date.now() + data.expires_in * 1000
-      }
+        expiresAt: Date.now() + data.expires_in * 1000,
+      };
     } catch (error) {
       await catcher<void>()
         .on(ApiError, (apiError) => {