+++ src/zendesk/model/zendesk-config.ts
@@ -1,7 +1,6 @@
 import { Config } from '../../config.js';
 
 export interface ZendeskConfig extends Config {
-  sourceUserAgent?: string;
   zendeskUsername?: string;
   zendeskPassword?: string;
   zendeskBaseUrl?: string;